Share dismissable-banner localStorage logic via a hook
NotificationBanner (merged from main) used the same setState-in-effect
localStorage pattern that react-hooks now rejects and Disclaimer already
had rewritten. Extract the useSyncExternalStore version into
lib/useDismissableBanner.js and use it from both banners.
